requirements.txt
python-docx==0.8.11
Pillow==10.0.0
matplotlib==3.7.2
numpy==1.24.3
openai==1.3.0
python-dotenv==1.0.0 